slightly changed UserController, Services

+++ src/main/java/cz/cvut/ear/clubevidence/service/CompetitionRecordService.java
@@ -1,77 +1,72 @@
 package cz.cvut.ear.clubevidence.service;
 
 import cz.cvut.ear.clubevidence.dao.CompetitionRecordDao;
+import cz.cvut.ear.clubevidence.exception.ExceptionGeneral;
+import cz.cvut.ear.clubevidence.exception.NotFoundException;
 import cz.cvut.ear.clubevidence.model.CompetitionRecord;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class CompetitionRecordService {
     private final CompetitionRecordDao competitionRecordDao;
 
-    /**
-     * Instantiates a new City service.
-     *
-     * @param competitionRecordDao the city dao
-     */
+
     @Autowired
     public CompetitionRecordService(CompetitionRecordDao competitionRecordDao) {
         this.competitionRecordDao = competitionRecordDao;
     }
 
-    /**
-     * Find all page.
-     *
-     // * @param pageable the pageable
-     // * @param name     the name
-     * @return the page
-     */
+
     @Transactional(readOnly = true)
     public List<CompetitionRecord> getAllCompetitionRecords() {
         return competitionRecordDao.findAll();
     }
 
-    /**
-     * Find city.
-     *
-     * @param id the id
-     * @return the city
-     */
+
     @Transactional(readOnly = true)
     public CompetitionRecord getCompetitionRecordById(Integer id) {
+        Objects.requireNonNull(id);
+        if(id<=0){
+            throw new IllegalArgumentException("Invalid id. It must be a positive integer.");
+        }
+        if(!competitionRecordDao.exists(id)){
+            throw new NotFoundException("Competition record not found");
+        }
         return competitionRecordDao.find(id);
     }
 
-    /**
-     * Persist.
-     *
-     //    * @param city the city
-     */
+
     @Transactional
     public void createCompetitionRecord(CompetitionRecord competitionRecord) {
+        Objects.requireNonNull(competitionRecord);
+        if(competitionRecordDao.exists(competitionRecord.getId())){
+            throw new ExceptionGeneral("Competition record already exists");
+        }
         competitionRecordDao.persist(competitionRecord);
     }
 
-    /**
-     * Update.
-     *
-     //  * @param city the city
-     */
+
     @Transactional
     public void updateCompetitionRecord(CompetitionRecord competitionRecord) {
+        Objects.requireNonNull(competitionRecord);
+        if(!competitionRecordDao.exists(competitionRecord.getId())){
+            throw new NotFoundException("Competition record not found");
+        }
         competitionRecordDao.update(competitionRecord);
     }
 
-    /**
-     * Remove.
-     *
-     // * @param city the city
-     */
+
     @Transactional
     public void deleteCompetitionRecord(CompetitionRecord competitionRecord) {
+        Objects.requireNonNull(competitionRecord);
+        if(!competitionRecordDao.exists(competitionRecord.getId())){
+            throw new NotFoundException("Competition record not found");
+        }
         competitionRecordDao.remove(competitionRecord);
     }
 }
